refactor: Remove "teacher" folder since students cannot edit a laboratory
