+++ src/screens/edit-laboratory/components/markdown-block/EditableMarkdownBlock.tsx
