+++ src/screens/edit-laboratory/components/markdown-block/MarkdownBlockDropDown.tsx
@@ -1,8 +1,10 @@
 import { EditLaboratoryContext } from "@/context/laboratories/EditLaboratoryContext";
 import { EditLaboratoryActionType } from "@/hooks/laboratories/editLaboratoryTypes";
-import { deleteTestBlockService } from "@/services/blocks/delete-test-block.service";
+import { deleteMarkdownBlockService } from "@/services/blocks/delete-markdown-block.service";
+import { updateMarkdownBlockContentService } from "@/services/blocks/update-markdown-block-content.service";
+import { MarkdownBlock } from "@/types/entities/laboratory-entities";
 import { ArrowDown, ArrowUp, MoreVertical, Save, Trash2 } from "lucide-react";
-import { RefObject, useContext } from "react";
+import { useContext } from "react";
 import { toast } from "sonner";
 
 import {
@@ -12,32 +14,41 @@ import {
   DropdownMenuLabel,
   DropdownMenuSeparator,
   DropdownMenuTrigger
-} from "../../../../../components/ui/dropdown-menu";
+} from "../../../../components/ui/dropdown-menu";
 
-interface TestBlockDropdown {
+interface MarkdownBlockDropDown {
   blockUUID: string;
   blockIndex: number;
-  formRef: RefObject<HTMLFormElement>;
 }
 
-export const TestBlockDropDown = ({
+export const MarkdownBlockDropDown = ({
   blockUUID,
-  blockIndex,
-  formRef
-}: TestBlockDropdown) => {
-  const { laboratoryStateDispatcher } = useContext(EditLaboratoryContext);
+  blockIndex
+}: MarkdownBlockDropDown) => {
+  const { laboratoryState, laboratoryStateDispatcher } = useContext(
+    EditLaboratoryContext
+  );
+  const { laboratory } = laboratoryState;
 
-  const handleSaveTestBlock = async () => {
-    formRef.current?.dispatchEvent(
-      new Event("submit", {
-        cancelable: true,
-        bubbles: true
-      })
-    );
+  const block = laboratory?.blocks.find(
+    (b) => b.uuid === blockUUID
+  ) as MarkdownBlock;
+
+  const handleSaveMarkdownBlock = async () => {
+    const { success, message } = await updateMarkdownBlockContentService({
+      markdownBlockUUID: blockUUID,
+      content: block.content
+    });
+
+    if (!success) {
+      toast.error(message);
+    } else {
+      toast.success(message);
+    }
   };
 
-  const handleDeleteTestBlock = async () => {
-    const { success, message } = await deleteTestBlockService(blockUUID);
+  const handleDeleteMarkdownBlock = async () => {
+    const { success, message } = await deleteMarkdownBlockService(blockUUID);
     if (!success) {
       toast.error(message);
     }
@@ -72,11 +83,11 @@ export const TestBlockDropDown = ({
           <ArrowDown className="mr-2 aspect-square h-5" />
           Move down
         </DropdownMenuItem>
-        <DropdownMenuItem onClick={handleSaveTestBlock}>
+        <DropdownMenuItem onClick={handleSaveMarkdownBlock}>
           <Save className="mr-2 aspect-square h-5" />
           Save changes
         </DropdownMenuItem>
-        <DropdownMenuItem onClick={handleDeleteTestBlock}>
+        <DropdownMenuItem onClick={handleDeleteMarkdownBlock}>
           <Trash2 className="mr-2 aspect-square h-5" />
           Delete block
         </DropdownMenuItem>
