+++ src/screens/edit-laboratory/components/test-block/TestBlockDropDown.tsx
@@ -1,10 +1,8 @@
 import { EditLaboratoryContext } from "@/context/laboratories/EditLaboratoryContext";
 import { EditLaboratoryActionType } from "@/hooks/laboratories/editLaboratoryTypes";
-import { deleteMarkdownBlockService } from "@/services/blocks/delete-markdown-block.service";
-import { updateMarkdownBlockContentService } from "@/services/blocks/update-markdown-block-content.service";
-import { MarkdownBlock } from "@/types/entities/laboratory-entities";
+import { deleteTestBlockService } from "@/services/blocks/delete-test-block.service";
 import { ArrowDown, ArrowUp, MoreVertical, Save, Trash2 } from "lucide-react";
-import { useContext } from "react";
+import { RefObject, useContext } from "react";
 import { toast } from "sonner";
 
 import {
@@ -14,41 +12,32 @@ import {
   DropdownMenuLabel,
   DropdownMenuSeparator,
   DropdownMenuTrigger
-} from "../../../../../components/ui/dropdown-menu";
+} from "../../../../components/ui/dropdown-menu";
 
-interface MarkdownBlockDropDown {
+interface TestBlockDropdown {
   blockUUID: string;
   blockIndex: number;
+  formRef: RefObject<HTMLFormElement>;
 }
 
-export const MarkdownBlockDropDown = ({
+export const TestBlockDropDown = ({
   blockUUID,
-  blockIndex
-}: MarkdownBlockDropDown) => {
-  const { laboratoryState, laboratoryStateDispatcher } = useContext(
-    EditLaboratoryContext
-  );
-  const { laboratory } = laboratoryState;
+  blockIndex,
+  formRef
+}: TestBlockDropdown) => {
+  const { laboratoryStateDispatcher } = useContext(EditLaboratoryContext);
 
-  const block = laboratory?.blocks.find(
-    (b) => b.uuid === blockUUID
-  ) as MarkdownBlock;
-
-  const handleSaveMarkdownBlock = async () => {
-    const { success, message } = await updateMarkdownBlockContentService({
-      markdownBlockUUID: blockUUID,
-      content: block.content
-    });
-
-    if (!success) {
-      toast.error(message);
-    } else {
-      toast.success(message);
-    }
+  const handleSaveTestBlock = async () => {
+    formRef.current?.dispatchEvent(
+      new Event("submit", {
+        cancelable: true,
+        bubbles: true
+      })
+    );
   };
 
-  const handleDeleteMarkdownBlock = async () => {
-    const { success, message } = await deleteMarkdownBlockService(blockUUID);
+  const handleDeleteTestBlock = async () => {
+    const { success, message } = await deleteTestBlockService(blockUUID);
     if (!success) {
       toast.error(message);
     }
@@ -83,11 +72,11 @@ export const MarkdownBlockDropDown = ({
           <ArrowDown className="mr-2 aspect-square h-5" />
           Move down
         </DropdownMenuItem>
-        <DropdownMenuItem onClick={handleSaveMarkdownBlock}>
+        <DropdownMenuItem onClick={handleSaveTestBlock}>
           <Save className="mr-2 aspect-square h-5" />
           Save changes
         </DropdownMenuItem>
-        <DropdownMenuItem onClick={handleDeleteMarkdownBlock}>
+        <DropdownMenuItem onClick={handleDeleteTestBlock}>
           <Trash2 className="mr-2 aspect-square h-5" />
           Delete block
         </DropdownMenuItem>
