+++ src/screens/edit-laboratory/skeletons/LaboratoryBlockSkeleton.tsx
@@ -1,5 +1,5 @@
 import { Skeleton } from "@/components/ui/skeleton";
 
 export const LaboratoryBlockSkeleton = () => {
-  return <Skeleton className="h-96 w-full space-y-4" />;
+  return <Skeleton className="mb-4 h-96 w-full" />;
 };
