+++ src/screens/edit-laboratory/skeletons/LaboratoryDetailsSkeleton.tsx
@@ -1,5 +1,5 @@
 import { Skeleton } from "@/components/ui/skeleton";
 
 export const LaboratoryDetailsSkeleton = () => {
-  return <Skeleton className="h-40 w-full space-y-4" />;
+  return <Skeleton className="mb-4 h-40 w-full" />;
 };
